Renamed important files, patched README for current project status.
